[chore] optimize imports
