+++ Sources/ElasticSwift/Errors/Errors.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import ElasticSwiftCore
-import ElasticSwiftNetworking
 
 //MARK:- ERRORS
