+++ Sources/ElasticSwift/Requests/IndicesRequests.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import Logging
 import NIOHTTP1
 import ElasticSwiftCore
 import ElasticSwiftCodableUtils
