+++ Sources/ElasticSwiftNetworking/ElasticSwiftNetworking.swift
@@ -8,6 +8,22 @@ import NIOSSL
 #endif
 import ElasticSwiftCore
 
+//MARK:- Timeouts
+
+public class Timeouts {
+    
+    public static let DEFAULT_TIMEOUTS: Timeouts = Timeouts(read: TimeAmount.milliseconds(1000), connect: TimeAmount.milliseconds(3000))
+    
+    public let read: TimeAmount?
+    public let connect: TimeAmount?
+    
+    public init(read: TimeAmount? = nil, connect: TimeAmount? = nil) {
+        self.read = read
+        self.connect = connect
+    }
+    
+}
+
 //MARK:- HTTPAdaptorConfiguration
 
 /// Class holding HTTPAdaptor Config
